avoid warning under setopt warn_nested_var
Previously, the option would be forcibly unset. This simply avoids
triggering warning, allowing the option to be maintained.

In earlier branches, before the option had been unset (or avoided), a
shell running with `setopt warn_nested_var`, the following warning would
be produced:
> ```
> _zsh_highlight_main__precmd_hook:1: array parameter _zsh_highlight_main__command_type_cache set in enclosing scope in function _zsh_highlight_main__precmd_hook
> ```

The documentation for `WARN_NESTED_VAR` is:
> Print a warning message when an existing parameter from an enclosing
> function scope, or  global, is set  in a function by an assignment or in
> math context. […] Where possible, use of `typeset -g` to set the parame-
> ter suppresses the error, but note that this needs to be used every time
> the parameter is  set.